+++ web/src/main/java/com/homeapp/javatraining/config/MinioConfig.java
@@ -1,0 +1,28 @@
+package com.homeapp.javatraining.config;
+
+import io.minio.MinioClient;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConfigurationProperties(prefix = "minio")
+@Setter
+@Getter
+public class MinioConfig {
+
+    private String endpoint;
+    private String accessKey;
+    private String secretKey;
+    private String bucket;
+
+    @Bean
+    public MinioClient minioClient() {
+        return MinioClient.builder()
+                .endpoint(endpoint)
+                .credentials(accessKey, secretKey)
+                .build();
+    }
+}